% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_df1.R
\docType{data}
\name{dep_df1}
\alias{dep_df1}
\title{Testing Scores of 100 Subjects from Rater/Time 1 in Dependent Case}
\format{
A dataframe with 100 rows and 4 columns
}
\usage{
dep_df1
}
\description{
A dataset containing scores for 100 subjects, where each subject has 4
trials. This is to be used in an example for the case of
comparing 2 dependent ICCs (function dep_ci()), where this dataset
contains the scores measured by Rater 1 (or at time 1) of the 100 subjects
and dataset dep_df2 contains the scores measured by Rater 2 (or at time
2) of the same 100 subjects.
}
\keyword{datasets}
