% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_ex_df1.R
\docType{data}
\name{dep_ex_df1}
\alias{dep_ex_df1}
\title{Testing Scores of 100 Subjects from Rater/Time 1 in Dependent Case}
\format{
A dataframe with 100 rows and 4 columns
}
\usage{
dep_ex_df1
}
\description{
A dataset containing scores with 4 trials for 100 subjects taken by
the first rater/time. This is to be used in an example for the case of
comparing 2 dependent ICCs (function dep_ci()), where this dataset
contains the scores measured by Rater/Time 1 of the 100 subjects and dataset
dep_ex_df2 contains the scores measured by Rater/Time 2 of the same 100 subjects.
}
\keyword{datasets}
