% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_df2.R
\docType{data}
\name{indep_df2}
\alias{indep_df2}
\title{Testing Scores of 80 Subjects from Rater/Time 2 in Independent Case}
\format{
A dataframe with 80 rows and 4 columns
}
\usage{
indep_df2
}
\description{
A dataset containing scores for 80 subjects, where each subject has 4
trials. This is to be used in an example for the case of
comparing 2 independent ICCs (function indep_ci()), where this dataset
contains the scores measured by Rater 2 (or at time 2) of the 80 subjects
and dataset indep_df1 contains the scores measured by Rater 1 (or at time
\enumerate{
\item of 100 different subjects.
}
}
\keyword{datasets}
