% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_ex_df1.R
\docType{data}
\name{indep_ex_df1}
\alias{indep_ex_df1}
\title{Testing Scores of 100 Subjects from Rater/Time 1 in Independent Case}
\format{
A dataframe with 100 rows and 4 columns
}
\usage{
indep_ex_df1
}
\description{
A dataset containing scores with 4 trials for 100 subjects taken by
the first rater/time. This is to be used in an example for the case of
comparing 2 independent ICCs (function indep_ci()), where this dataset
contains the scores measured by Rater/Time 1 of 100 subjects and dataset
indep_ex_df2 contains the scores measured by Rater/Time 2 of a different
set of 80 subjects.
}
\keyword{datasets}
