% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_ex_df2.R
\docType{data}
\name{dep_ex_df2}
\alias{dep_ex_df2}
\title{Testing Scores of 100 Subjects from Rater 2 in Dependent Case}
\format{
A dataframe with 100 rows and 4 columns
}
\usage{
dep_ex_df2
}
\description{
A dataset containing scores with 4 trials for 100 subjects taken by
the second rater. This is to be used in an example for the case of
comparing 2 dependent ICCs (function dep_ci()), where this dataset
contains the scores measured by Rater 2 of the 100 subjects and dataset
dep_ex_df1 contains the scores measured by Rater 1 of the same 100 subjects.
}
\keyword{datasets}
